Multiple changes to web configuration interface.
Adds:
* Uptime reporting,
* AP security settings.
* Debug information enable/disable.

+++ firmware/include/long_millis.h
@@ -4,12 +4,12 @@
 #include <SmingCore/SmingCore.h>
 
 
-static uint32_t millis_offset = 0;
-static uint32_t micros_extra = 0;
+static uint64_t millis_offset = 0;
+static uint64_t micros_extra = 0;
 static uint32_t last_us = 0;
 
 
-uint32_t long_millis() {
+uint64_t long_millis() {
    uint32_t us = micros();
    if (us < last_us) {
       millis_offset += 4294967;
